pos Comprar finish

+++ app/src/main/java/com/carvalho/pi/EditProdutoFragment.kt
@@ -1,5 +1,6 @@
 package com.carvalho.pi
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -40,9 +41,22 @@ class EditProdutoFragment : Fragment() {
         }
 
         binding.btnDeletarEdit.setOnClickListener {
-            viewModel.deleteProduto(viewModel.produtoSelecionado!!.id)
-            findNavController().navigate(R.id.action_editProdutoFragment_to_listagemFragment)
-            Toast.makeText(context, "Produto Foi Deletado com Sucesso", Toast.LENGTH_SHORT).show()
+
+            val builder = AlertDialog.Builder(context)
+            builder.setTitle("Delete")
+            builder.setMessage("Deseja realmente deletar esse produto")
+
+            builder.setPositiveButton("Deletar") { _, _ ->
+                viewModel.deleteProduto(viewModel.produtoSelecionado!!.id)
+                findNavController().navigate(R.id.action_editProdutoFragment_to_listagemFragment)
+                Toast.makeText(context, "Produto Foi Deletado com Sucesso", Toast.LENGTH_SHORT).show()
+            }
+
+            builder.setNegativeButton("Cancelar") { _, _  ->
+                Toast.makeText(context, "Produto Não Foi Deletado", Toast.LENGTH_SHORT).show()
+            }
+
+            builder.show()
         }
 
         binding.btnEditar.setOnClickListener {
